requirements.txt
aiohttp==3.7.4.post0
alpha-vantage==2.3.1
altgraph==0.17
appdirs==1.4.4
appnope==0.1.2
arabic-reshaper==2.1.3
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
auto-py-to-exe==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
bootstrap4==0.1.0
bottle==0.12.19
bottle-websocket==0.2.9
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-v5==1.0.4
django-bootstrap4==3.0.0
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-datetimepicker==3.14
django-filter==2.4.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
django-phone-field==1.8.1
django-richtextfield==1.6.1
djangorestframework==3.12.4
Eel==0.12.4
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
fusepy==3.0.1
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
install==1.3.4
intrinio-sdk==5.11.0
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.6.3
macholib==1.14
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
multidict==5.1.0
multitasking==0.0.9
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
packaging==21.0
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2020.11.15
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyJWT==1.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.18.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
pyzmq==22.1.0
reportlab==3.5.67
requests==2.25.1
retrying==1.3.3
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.7.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.18
sqlparse==0.4.1
terminado==0.10.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
twilio==6.59.1
typing-extensions==3.10.0.0
urllib3==1.26.4
virtualenv==20.4.3
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
whichcraft==0.6.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
xhtml2pdf==0.2.5
yahoo-finance==1.4.0
yarl==1.6.3
yfinance==0.1.59
youtube-dl==2021.5.16
ytfs==0.99a5
zope.event==4.5.0
zope.interface==5.4.0
